+++ data-structures/src/main/java/za/co/oldtimerza/data/Dequeue.java
@@ -1,0 +1,5 @@
+package za.co.oldtimerza.data;
+
+public class Dequeue extends LinkedList
+{
+}
