+++ data-structures/src/main/java/za/co/oldtimerza/data/List.java
@@ -1,0 +1,11 @@
+package za.co.oldtimerza.data;
+
+public interface List{
+    void add(LinkableContainer node);
+    void add(int index, LinkableContainer node);
+    void remove(int node);
+    LinkableContainer getFirst();
+    LinkableContainer getLast();
+    LinkableContainer get(int index);
+
+}
