+++ data-structures/src/test/java/za/co/oldtimerza/data/LinkedListTest.java
@@ -4,16 +4,16 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class DoublyLinkedListTest {
+public class LinkedListTest{
 
-    private DoublyLinkedList list;
+    private LinkedList list;
     private LinkableContainer node1;
     private LinkableContainer node2;
     private LinkableContainer node3;
 
     @Before
-    public void setup(){
-        this.list = new DoublyLinkedList();
+   public void setup(){
+        this.list = new LinkedList();
         this.node1 = new Node<String>("node1");
         this.node2 = new Node<String>("node2");
         this.node3 = new Node<String>("node3");
@@ -25,27 +25,27 @@ public class DoublyLinkedListTest {
     @Test
     public void shouldAddHeadToList()
     {
-        Assert.assertEquals(this.list.getHead(),node1);
+        Assert.assertEquals(this.list.getFirst(),node1);
     }
 
     @Test
     public void shouldAddNodeToEndOfList()
     {
-        Assert.assertEquals(this.list.getTail(),node3);
+        Assert.assertEquals(this.list.getLast(),node3);
     }
 
     @Test
     public void shoudlAddNodeAtPosition()
     {
     LinkableContainer node4 = new Node<Integer>();
-    this.list.addAtPosition(1,node4);
+    this.list.add(1,node4);
     Assert.assertEquals(this.list.get(2).getNext(), node3);
     }
 
     @Test
     public void shouldGetListLength()
     {
-        Assert.assertEquals(this.list.getNumberOfElements(),3);
+        Assert.assertEquals(this.list.length(),3);
     }
 
     @Test
